Return normalized base64 for encryptedPersonalId and move BlindRequest to models
