+++ app/services/oprf/oprf_service.py
@@ -1,29 +1,12 @@
 import base64
 import pyoprf
 from jwcrypto import jwk
-from pydantic import BaseModel, Field, field_validator
 
 from app.services.oprf.jwe_token import BlindJwe
+from app.models.requests import BlindRequest
 import logging
 
 logger = logging.getLogger(__name__)
-
-
-class BlindRequest(BaseModel):
-    encryptedPersonalId: str = Field(..., min_length=2)
-    recipientOrganization: str = Field(..., min_length=2)
-    recipientScope: str = Field(..., min_length=2)
-
-    @field_validator("encryptedPersonalId")
-    def validate_base64(cls, v: str) -> str:
-        try:
-            pad = "=" * ((4 - len(v) % 4) % 4)
-            base64.urlsafe_b64decode(v + pad)
-        except Exception as e:
-            logger.error("invalid base64url string: %r", v)
-            raise ValueError(f"must be base64url: {e}")
-
-        return v
 
 
 class OprfService:
